fix: remaining mypy errors

+++ pyshocks/tools.py
@@ -173,7 +173,7 @@ def estimate_order_of_convergence(x: Array, y: Array) -> tuple[Scalar, Scalar]:
     if x.size <= 1:
         raise RuntimeError("Need at least two values to estimate order.")
 
-    eps = jnp.finfo(x.dtype).eps
+    eps = jnp.finfo(x.dtype).eps  # type: ignore[no-untyped-call]
     c = jnp.polyfit(jnp.log10(x + eps), jnp.log10(y + eps), 1)
     return 10 ** c[-1], c[-2]
 
@@ -272,7 +272,7 @@ class EOCRecorder:
 
         _, error = self._history
         if atol is None:
-            atol = 1.0e2 * jnp.finfo(error.dtype).eps
+            atol = 1.0e2 * jnp.finfo(error.dtype).eps  # type: ignore[no-untyped-call]
 
         return bool(self.estimated_order >= (order - slack) or jnp.max(error) < atol)
 
@@ -877,7 +877,7 @@ def save_animation(
     if legends is not None:
         ax.set_ylabel(ylabel)
     ax.set_xlim((float(x[0]), float(x[-1])))
-    ax.set_ylim((ymin - 0.1 * jnp.abs(ymin), ymax + 0.1 * jnp.abs(ymax)))
+    ax.set_ylim((float(ymin - 0.1 * jnp.abs(ymin)), float(ymax + 0.1 * jnp.abs(ymax))))
     ax.margins(0.05)
 
     if legends:
